Rename fragments package to fragment
