+++ android/app/src/main/kotlin/net/mullvad/mullvadvpn/ui/fragment/FragmentArgumentConstant.kt
@@ -1,3 +1,3 @@
-package net.mullvad.mullvadvpn.ui.fragments
+package net.mullvad.mullvadvpn.ui.fragment
 
 const val ACCOUNT_TOKEN_ARGUMENT_KEY = "accountToken"
